Fix two test failures caused by Windows mangling of function types.
On Windows the function template `template <class T> void test()` has
the same mangled name when instantiated with the distinct types `void()`
and `void() noexcept`. When this occurs Clang emits an error. This error
was causing two type-traits tests to fail.

However this can be worked around by using class templates instead of
function templates, which is what this patch does to fix the errors.

git-svn-id: http://llvm.org/svn/llvm-project/libcxx/trunk@302380 91177308-0d34-0410-b5e6-96231b3b80d8