[libc++] (Take 2) Add a test that uses the debug database from multiple threads
In r358591, I added a test that uses the debug database from multiple
threads and that helped us uncover the problem that was fixed in r355367.
However, the test broke the tsan CI bots, and I think the problem is the
test allocator that was used in the test (which is not thread safe).

I'm committing again without using the test allocator, and in a separate
test file.

git-svn-id: http://llvm.org/svn/llvm-project/libcxx/trunk@358610 91177308-0d34-0410-b5e6-96231b3b80d8