[libc++] Use a function to set warning flags per target
This is part of a larger shift to move to per-target settings and
eradicate global variables from the CMake build. I'm starting small
with warnings only because those are easy to transition over and I
want to see how it pans out, but we can handle all flags like exceptions
and RTTI in the future.

git-svn-id: http://llvm.org/svn/llvm-project/libcxx/trunk@373511 91177308-0d34-0410-b5e6-96231b3b80d8